- docs: add api docs for auth - refactor: split function auth per file - chore: add field role on table user
