+++ index.js
@@ -9,10 +9,6 @@ app.use(express.json());
 app.use(cors());
 app.use(authRouter);
 
-app.use("/ping", (req, res) => {
-   res.send('halo')
-})
-
 app.listen(config.APP_PORT, () => {
    console.log(`Server Activated On Port ${config.APP_PORT}`);
 });
