+++ prisma/schema.prisma
@@ -23,6 +23,7 @@ model User {
    username   String   @unique
    email      String   @unique
    password   String
+   role       String
    created_at DateTime @default(now())
    jwt        Jwt[]
 }
